backend/requirements.txt
fastapi
uvicorn
supabase
python-dotenv
pydantic
google-generative-ai
